still could not meet performance requirements
planned to do some profiling but failed, and could not guess which part
slows down the program.

the problem is related to large integer factorization, or probably.
expect an integer taking about 10~20s to run and profile, but this is
hard to reach. some got factorized in a blink of an eye but others were
slow enough to receive a SIGKILL. this difference depends on the
magnitudes of factors, with integers made up of small prime factors
collapse faster than those of large factors. probably try to compose
integers of proper factors (about 1e4 ~ 1e5).